refactor: mutualize endpoint and entrypoint connectors factory
