+++ src/main/java/io/gravitee/gateway/jupiter/api/connector/entrypoint/EntrypointConnectorConfiguration.java
@@ -13,9 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.gateway.jupiter.api.entrypoint.async;
+package io.gravitee.gateway.jupiter.api.connector.entrypoint;
 
-import io.gravitee.gateway.jupiter.api.context.MessageExecutionContext;
-import io.gravitee.gateway.jupiter.api.entrypoint.EntrypointConnector;
-
-public interface EntrypointAsyncConnector extends EntrypointConnector<MessageExecutionContext> {}
+/**
+ * @author Guillaume LAMIRAND (guillaume.lamirand at graviteesource.com)
+ * @author GraviteeSource Team
+ */
+public interface EntrypointConnectorConfiguration {}
