--- src/main/java/io/gravitee/gateway/jupiter/api/endpoint/EndpointConnector.java
@@ -1,30 +1,0 @@
-/**
- * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.gravitee.gateway.jupiter.api.endpoint;
-
-import io.gravitee.gateway.jupiter.api.ApiType;
-import io.gravitee.gateway.jupiter.api.ConnectorMode;
-import io.gravitee.gateway.jupiter.api.context.ExecutionContext;
-import io.reactivex.Completable;
-import java.util.Set;
-
-public interface EndpointConnector<T extends ExecutionContext> {
-    ApiType supportedApi();
-
-    Set<ConnectorMode> supportedModes();
-
-    Completable connect(final T executionContext);
-}
